+++ resources/views/colaboradores.blade.php
@@ -1,117 +1,186 @@
-@extends('layouts.app')
+@extends('adminlte::page')
 
-@section('title_page')
-    <!-- Content Header (Page header) -->
-    <section class="content-header">
+@section('title', 'Usuários')
+
+@section('content_header')
         <h1>Colaboradores</h1>
         <ol class="breadcrumb">
             <li><a href="colaboradores"><i class="fa  fa-male"></i> Colaboradores</a></li>
-
         </ol>
-    </section>
-@endsection
-@section('content_page')
+@stop
 
-    @include('layouts.sidebar')
-
-    <!-- Main content -->
-    <section class="content">
+@section('content')
         <div class="row">
 
-            <div class="col-md-12">
-                <div class="box box-primary">
-                    <div class="box-header">
-                        <div class="box-header with-border">
+                @if(Request::is('*/editar'))
+                @can('edit_colaborador')
+                    <div class="col-md-12">
+                        <div class="box box-primary">
+                            <div class="box-header">
+                                <div class="box-header with-border">
+                                        <i class="fa fa-pencil-square-o"></i>       <h3 class="box-title">Atualizar</h3>
+                                </div>
 
-                            @if(Request::is('*/editar'))
-                                <i class="fa fa-pencil-square-o"></i>       <h3 class="box-title">Atualizar</h3>
-                            @else
-                                <i class="fa fa-user-plus"></i>   <h3 class="box-title">Cadastrar Colaborador</h3>
-                            @endif
-                        </div>
+                                    {!! Form::model($edit,['method' => 'PATCH','class' => 'form-horizontal', 'url' => 'colaboradores/'.$edit->id]) !!}
 
-                        @if(Request::is('*/editar'))
-                            {!! Form::model($edit,['method' => 'PATCH','class' => 'form-horizontal', 'url' => 'colaboradores/'.$edit->id]) !!}
-                        @else
-                            {!! Form::open(array('route' => 'colaboradores.store', 'class' => 'form-horizontal')) !!}
-                        @endif
+                                <div class="fetched-data">
+                                    {{ csrf_field() }}
+                                    <div class="box-body">
 
-                        <div class="fetched-data">
-                            {{ csrf_field() }}
-                            <div class="box-body">
+                                        <!-- Mensagem de Ação -->
+                                        @if(Session::has('mensagem_sucesso'))
+                                            <div class="callout callout-success">
+                                                {{ Session::get('mensagem_sucesso') }}
+                                            </div>
+                                        @elseif(Session::has('mensagem_update'))
+                                            <div class="callout callout-warning">
+                                                {{ Session::get('mensagem_update') }}
+                                            </div>
+                                        @elseif(Session::has('mensagem_destroy'))
+                                            <div class="callout callout-danger">
+                                                {{ Session::get('mensagem_destroy') }}
+                                            </div>
+                                        @endif
 
-                                <!-- Mensagem de Ação -->
-                                @if(Session::has('mensagem_sucesso'))
-                                    <div class="callout callout-success">
-                                        {{ Session::get('mensagem_sucesso') }}
-                                    </div>
-                                @elseif(Session::has('mensagem_update'))
-                                    <div class="callout callout-warning">
-                                        {{ Session::get('mensagem_update') }}
-                                    </div>
-                                 @elseif(Session::has('mensagem_destroy'))
-                                        <div class="callout callout-danger">
-                                            {{ Session::get('mensagem_destroy') }}
+                                        <div class="form-group{{ $errors->has('nome') ? ' has-error' : '' }}">
+                                            {!!  Form::label('nome','Nome',array('class' => 'col-sm-2 control-label')) !!}
+                                            <div class="col-sm-6">
+                                                {!! Form::input('text','nome',old('nome'),array('class' => 'form-control','placeholder' => 'Nome do colaborador'))!!}
+                                                @if ($errors->has('nome'))
+                                                    <span class="help-block"><strong>{{ $errors->first('nome') }}</strong></span>
+                                                @endif
+                                            </div>
                                         </div>
-                                @endif
 
-                                <div class="form-group{{ $errors->has('nome') ? ' has-error' : '' }}">
-                                    {!!  Form::label('nome','Nome',array('class' => 'col-sm-2 control-label')) !!}
-                                    <div class="col-sm-6">
-                                        {!! Form::input('text','nome',old('nome'),array('class' => 'form-control','placeholder' => 'Nome do colaborador'))!!}
-                                        @if ($errors->has('nome'))
-                                            <span class="help-block"><strong>{{ $errors->first('nome') }}</strong></span>
-                                        @endif
+                                        <div class="form-group{{ $errors->has('cargo') ? ' has-error' : '' }}">
+                                            {!!  Form::label('cargo','Cargo',array('class' => 'col-sm-2 control-label')) !!}
+                                            <div class="col-sm-6">
+                                                {{ Form::select('cargo' ,[
+                                                                                'professor' => 'Professor',
+                                                                                'administrativo' => 'Administrativo',
+                                                                                'outros' => 'Outros',
+
+                                                                                ],'Professor',['class' => 'form-control'])
+                                                                             }}
+
+                                                @if ($errors->has('cargo'))
+                                                    <span class="help-block"><strong>{{ $errors->first('cargo') }}</strong></span>
+                                                @endif
+                                            </div>
+                                        </div>
+
+                                        <div class="box-footer">
+                                                <a class="btn btn-default btn-close" href="{{ route('colaboradores.index') }}">Cancelar</a>
+                                                {!! Form::submit('Atualziar', ['class' => 'btn btn-info' ]) !!}
+                                        </div>
+                                        {!! @Form::close() !!}
+                                    </div><!-- /.modal-content -->
+                                </div><!-- /.modal-dialog -->
+                            </div><!-- /.modal -->
+                        </div>
+                        @endcan
+
+
+                        @else
+                            @can('store_colaborador')
+                                <div class="col-md-12">
+                                    <div class="box box-primary">
+                                        <div class="box-header">
+                                            <div class="box-header with-border">
+
+                                                @if(Request::is('*/editar'))
+                                                    <i class="fa fa-pencil-square-o"></i>       <h3 class="box-title">Atualizar</h3>
+                                                @else
+                                                    <i class="fa fa-user-plus"></i>   <h3 class="box-title">Cadastrar Colaborador</h3>
+                                                @endif
+                                            </div>
+
+                                            @if(Request::is('*/editar'))
+                                                {!! Form::model($edit,['method' => 'PATCH','class' => 'form-horizontal', 'url' => 'colaboradores/'.$edit->id]) !!}
+                                            @else
+                                                {!! Form::open(array('route' => 'colaboradores.store', 'class' => 'form-horizontal')) !!}
+                                            @endif
+
+                                            <div class="fetched-data">
+                                                {{ csrf_field() }}
+                                                <div class="box-body">
+
+                                                    <!-- Mensagem de Ação -->
+                                                    @if(Session::has('mensagem_sucesso'))
+                                                        <div class="callout callout-success">
+                                                            {{ Session::get('mensagem_sucesso') }}
+                                                        </div>
+                                                    @elseif(Session::has('mensagem_update'))
+                                                        <div class="callout callout-warning">
+                                                            {{ Session::get('mensagem_update') }}
+                                                        </div>
+                                                    @elseif(Session::has('mensagem_destroy'))
+                                                        <div class="callout callout-danger">
+                                                            {{ Session::get('mensagem_destroy') }}
+                                                        </div>
+                                                    @endif
+
+                                                    <div class="form-group{{ $errors->has('nome') ? ' has-error' : '' }}">
+                                                        {!!  Form::label('nome','Nome',array('class' => 'col-sm-2 control-label')) !!}
+                                                        <div class="col-sm-6">
+                                                            {!! Form::input('text','nome',old('nome'),array('class' => 'form-control','placeholder' => 'Nome do colaborador'))!!}
+                                                            @if ($errors->has('nome'))
+                                                                <span class="help-block"><strong>{{ $errors->first('nome') }}</strong></span>
+                                                            @endif
+                                                        </div>
+                                                    </div>
+
+                                                    <div class="form-group{{ $errors->has('cargo') ? ' has-error' : '' }}">
+                                                        {!!  Form::label('cargo','Cargo',array('class' => 'col-sm-2 control-label')) !!}
+                                                        <div class="col-sm-6">
+                                                            {{ Form::select('cargo' ,[
+                                                                                            'professor' => 'Professor',
+                                                                                            'administrativo' => 'Administrativo',
+                                                                                            'outros' => 'Outros',
+
+                                                                                            ],'Professor',['class' => 'form-control'])
+                                                                                         }}
+
+                                                            @if ($errors->has('cargo'))
+                                                                <span class="help-block"><strong>{{ $errors->first('cargo') }}</strong></span>
+                                                            @endif
+                                                        </div>
+                                                    </div>
+
+                                                    <div class="box-footer">
+                                                        @if(Request::is('*/editar'))
+                                                            <a class="btn btn-default btn-close" href="{{ route('colaboradores.index') }}">Cancelar</a>
+                                                            {!! Form::submit('Atualziar', ['class' => 'btn btn-info' ]) !!}
+                                                        @else
+                                                            {!! Form::submit('Cadastrar', ['class' => 'btn btn-info' ]) !!}
+                                                        @endif
+                                                    </div>
+                                                    {!! @Form::close() !!}
+                                                </div><!-- /.modal-content -->
+                                            </div><!-- /.modal-dialog -->
+                                        </div><!-- /.modal -->
                                     </div>
-                                </div>
 
-                                <div class="form-group{{ $errors->has('cargo') ? ' has-error' : '' }}">
-                                    {!!  Form::label('cargo','Cargo',array('class' => 'col-sm-2 control-label')) !!}
-                                    <div class="col-sm-6">
-                                        {{ Form::select('cargo' ,[
-                                                                        'professor' => 'Professor',
-                                                                        'administrativo' => 'Administrativo',
-                                                                        'outros' => 'Outros',
-
-                                                                        ],'Professor',['class' => 'form-control'])
-                                                                     }}
-
-                                        @if ($errors->has('cargo'))
-                                            <span class="help-block"><strong>{{ $errors->first('cargo') }}</strong></span>
-                                        @endif
-                                    </div>
-                                </div>
-
-                                <div class="box-footer">
-                                    @if(Request::is('*/editar'))
-                                        <a class="btn btn-default btn-close" href="{{ route('colaboradores.index') }}">Cancelar</a>
-                                        {!! Form::submit('Atualziar', ['class' => 'btn btn-info' ]) !!}
-                                    @else
-                                        {!! Form::submit('Cadastrar', ['class' => 'btn btn-info' ]) !!}
+                                    @endcan
                                     @endif
-                                </div>
-                                {!! @Form::close() !!}
-                            </div><!-- /.modal-content -->
-                        </div><!-- /.modal-dialog -->
-                    </div><!-- /.modal -->
-                </div>
 
 
 
-                <div class="box box-primary">
+                <div class="col-md-12">
+                            <div class="box box-primary">
                     <div class="box-header">
                         <i class="ion ion-clipboard"></i>
                         <h3 class="box-title">Lista de Colaboradores</h3>
                     </div>
                     <!-- /.box-header -->
                     <div class="box-body">
-                        <table id="example1" class="table table-bordered table-striped">
+                        <table id="example1" class="table table-striped table-bordered" cellspacing="0" width="100%">
                             <thead>
                             <tr>
                                 <th>Nome</th>
                                 <th>Cargo</th>
-                                <th>Ação</th>
-                                <th></th>
+                                @can('edit_colaborador') <th>Editar</th>  @endcan
+                                @can('delete_colaborador') <th>Exluir</th>  @endcan
                             </tr>
                             </thead>
                             <tbody>
@@ -119,19 +188,17 @@
                                 <tr>
                                     <td>{{$rows->nome}}</td>
                                     <td>{{$rows->cargo}}</td>
-                                    <td width="50"> <div class="tools">
 
-
+                                    @can('edit_colaborador') <td width="50"> <div class="tools">
                                             <a href="{{ route('colaboradores.edit', ['id' => $rows->id ]) }}" ><button  title="Editar" class="btn btn-warning"><i class="fa fa-edit"></i> Editar</button></a>
-
                                         </div>
-                                    </td>
-                                    <td width="50">
+                                    </td>@endcan
+                                    @can('delete_colaborador') <td width="50">
                                         {!! Form::open(['method'=>'DELETE', 'url' => '/colaboradores/'.$rows->id]) !!}
 
                                         <a><button  class="btn btn-danger" onclick="return confirm('Deseja remover o colaborador {{$rows->nome}}?')"  title="Excluir" ><i class="fa fa-trash-o"></i> Excluir</button></a>
                                         {!! Form::close() !!}
-                                    </td>
+                                    </td> @endcan
 
                                     @endforeach
                                 </tr>
@@ -152,21 +219,21 @@
             <!-- /.col -->
         </div>
         <!-- /.row -->
-    </section>
-    <!-- /.content -->
-    </div>
+        </div>
 
+@stop
 
-    </section>
-    <!-- /.box -->
+@section('css')
 
+@stop
 
+@section('js')
+    <script src="//code.jquery.com/jquery-1.12.4.js"></script>
+    <script src="https://cdn.datatables.net/1.10.16/js/jquery.dataTables.min.js"></script>
+    <script src="https://cdn.datatables.net/1.10.16/js/dataTables.bootstrap.min.js"></script>
 
+    <script>
 
-
-
-
-    <script type="text/javascript">
         $(function () {
             $("#example1").DataTable({      "oLanguage": {
                 "sProcessing":   "Processando...",
@@ -189,10 +256,4 @@
         });
     </script>
 
-@endsection
-
-
-@section('scripts')
-
-
-@endsection
+@stop
